Create TorrentDownloadedNotifier
